#125 Setup Core project to fetch BookShelves, Add BookShelfVM
