+++ ThePage/src/ThePage.Core/ThePage.Core.csproj
@@ -36,6 +36,7 @@
     <Folder Include="Utils\" />
     <Folder Include="ViewModels\Authentication\" />
     <Folder Include="Utils\MvvmCross\" />
+    <Folder Include="ViewModels\Bookshelf\" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\ThePage.Api\ThePage.Api.csproj" />
